Get minimal reate and edit working for volunteers
Everything is still very generic and string-based for the moment;
A volunteer-specific view will probably be the next step with a
prettier status (active/inactive) checkbox and orientation date
fields. However now that a database can be populated with test
volunteer entries, getting time records and tracking will be next.