Drawing the rest of the owl... Most core functionality implemented
Implemented categories, which are a lot like volunteers in a db sense,
having just a name and an active/inactive status.

Implemented sign in/out functionality (which needed volunteer and
category table manipulation in order to work).

Next step is to implement admin functionality including managing
volunteers and categories, as well as running reports. Most of the
code to do all this stuff is here, it's just a matter of organizing
it and cleaning out the stuff that isn't needed.